+++ packages/create-sitecore-jss/src/bin.test.ts
@@ -12,11 +12,11 @@ import * as initRunner from './init-runner';
 
 describe('bin', () => {
   describe('parseArgs', () => {
-    let originalArgv: string[];
+    let originalArgv: string[] = [];
 
     beforeEach(() => {
       // eslint-disable-next-line no-unused-labels
-      originalArgv: process.argv;
+      originalArgv = process.argv;
     });
 
     afterEach(() => {